requirements.txt
equinox==0.11.4
hydra-core==1.3.2
jax==0.4.23
matplotlib==3.8.3 
mph==1.2.3
numpy==1.26.4
omegaconf==2.3.0
optax==0.2.2
pandas==2.0.3
pytest==8.2.0
torch==2.2.1
tqdm==4.65.0
